@@ -11,9 +11,9 @@
     "prettier": "prettier --write 'api/**/src/**/*.ts' 'dashboard/src/**/*.ts'"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^3.7.0",
-    "@typescript-eslint/parser": "^3.7.0",
-    "eslint": "^7.5.0",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
+    "eslint": "^7.14.0",
     "eslint-plugin-prettier": "^3.1.4",
     "husky": "^4.2.3",
     "lerna": "^3.20.2",
