@@ -1,14 +1,20 @@
 import { Component, OnInit, AfterViewInit } from '@angular/core';
-import { FormArray, FormBuilder, FormGroup } from '@angular/forms';
-import { debounceTime, distinctUntilChanged, filter, switchMap, takeUntil, tap } from 'rxjs/operators';
-import { merge, of } from 'rxjs';
+import { FormBuilder, FormGroup, FormControl } from '@angular/forms';
+import { debounceTime, distinctUntilChanged, filter, takeUntil, tap, map } from 'rxjs/operators';
+import { merge } from 'rxjs';
 import { ToastrService } from 'ngx-toastr';
 
 import { CommonDataService } from '~/core/services/common-data.service';
 import { DestroyObservable } from '~/core/components/destroy-observable';
-import { Territory } from '~/core/entities/territory/territory';
+import { TerritoryTree, TerritoryLevelEnum } from '~/core/entities/territory/territory';
 import { OperatorVisilibityService } from '~/modules/operator/services/operator-visilibity.service';
 
+interface TerritoryVisibilityTree extends TerritoryTree {
+  control?: FormControl;
+  selected: boolean;
+}
+
+const isSelectableFilter = (t: TerritoryVisibilityTree) => t.activable || t.level === TerritoryLevelEnum.Towngroup;
 @Component({
   selector: 'app-operator-visibility-tree',
   templateUrl: './operator-visibility-tree.component.html',
@@ -18,11 +24,12 @@ export class OperatorVisibilityTreeComponent extends DestroyObservable implement
   searchFilter: FormGroup;
   checkAllValue = false;
 
-  territories: Territory[] = [];
+  territories: TerritoryVisibilityTree[] = [];
   territoryIdsToShow: number[] = [];
-  checkedTerritoryIds: number[] = [];
+  selectedTerritoryIds: number[] = [];
 
-  visibilityFormGroup: FormGroup;
+  visibilityFormControl: FormControl;
+  searchMode: boolean;
 
   constructor(
     private _commonDataService: CommonDataService,
@@ -33,34 +40,91 @@ export class OperatorVisibilityTreeComponent extends DestroyObservable implement
     super();
   }
 
+  swapCheck(ter: TerritoryVisibilityTree) {
+    console.log('> swapCheck', ter._id);
+    const selectedTerritoryIds = this.selectedTerritoryIds;
+    const selInd = selectedTerritoryIds.indexOf(ter._id);
+    const selected = selInd === -1;
+
+    if (selected) {
+      selectedTerritoryIds.push(ter._id);
+    } else {
+      selectedTerritoryIds.splice(selInd, 1);
+    }
+
+    this.visibilityFormControl.setValue(selectedTerritoryIds);
+  }
+
   ngOnInit(): void {
     this.initSearchForm();
     this.initVisibilityForm();
   }
 
+  flatTree(tree: Partial<TerritoryTree>[], indent = 0): TerritoryVisibilityTree[] {
+    const res = [];
+    if (tree)
+      for (const ter of tree) {
+        ter.indent = indent;
+        res.push(ter);
+        if (ter.children) {
+          const children = this.flatTree(ter.children, indent + 1);
+          for (const child of children) res.push({ ...child });
+        }
+      }
+
+    return res;
+  }
+
   ngAfterViewInit() {
     merge(
-      this._commonDataService.activableTerritories$.pipe(
+      this._commonDataService.territoriesTree$.pipe(
         filter((territories) => !!territories),
         tap((territories) => {
-          this.territories = territories;
+          this.territories = this.flatTree(territories);
         }),
       ),
       this.searchFilter.valueChanges.pipe(debounceTime(100)),
       this._operatorVisilibityService.operatorVisibility$.pipe(
         filter((territories) => !!territories),
-        tap((territoryIds) => (this.checkedTerritoryIds = territoryIds)),
+        tap((territoryIds) => (this.selectedTerritoryIds = territoryIds)),
       ),
     )
       .pipe(
         distinctUntilChanged(),
         debounceTime(100),
-        switchMap(() => {
-          const lowerCasedQuery = this.searchFilter ? this.searchFilter.controls.query.value.toLowerCase() : '';
-          const filteredTerritories = this.territories.filter((t) =>
-            `${t.name}`.toLowerCase().includes(lowerCasedQuery),
-          );
-          return of(filteredTerritories);
+        map(() => {
+          if (this.searchFilter && this.searchFilter.controls.query.value) {
+            this.searchMode = true;
+            const words = this.searchFilter.controls.query.value.toLowerCase().split(new RegExp('[ -_]'));
+            // score search based ordering
+            const territories = this.territories
+              // get only EPCI or AOM (activable)
+              .filter(isSelectableFilter)
+              // search for requested words and set score based on matching words amount
+              .map<TerritoryVisibilityTree & { score: number }>((t) => {
+                const nameLowerCase = t.name.toLowerCase();
+                const score = words.map((w) => nameLowerCase.split(w).length - 1).reduce((acc, score) => acc + score);
+                // return null for fast filter optimisation
+                return score > 0
+                  ? {
+                      ...t,
+                      score,
+                    }
+                  : null;
+              })
+              // filter found territory
+              .filter((t) => t !== null)
+              .sort((a, b) => {
+                if (a.score < b.score) return -1;
+                if (a.score > b.score) return 1;
+                return 0;
+              });
+            return territories;
+          } else {
+            this.searchMode = false;
+
+            return this.territories;
+          }
         }),
       )
       .pipe(takeUntil(this.destroy$))
@@ -72,10 +136,6 @@ export class OperatorVisibilityTreeComponent extends DestroyObservable implement
     this.loadVisibility();
   }
 
-  get territoriesFormArray(): FormArray {
-    return this.visibilityFormGroup.get('territories') as FormArray;
-  }
-
   get hasFilter(): boolean {
     return this.searchFilter.value;
   }
@@ -84,21 +144,8 @@ export class OperatorVisibilityTreeComponent extends DestroyObservable implement
     return this._operatorVisilibityService.isLoaded;
   }
 
-  public updateCheckAllCheckbox(): void {
-    this.checkAllValue = this.territories.length === this.checkedTerritoryIds.length;
-  }
-
   public save(): void {
-    const territoriesIds = (this.checkedTerritoryIds = this.territoriesFormArray.value.reduce(
-      (checkedTerritories: number[], checked: boolean, index: number) => {
-        if (checked) {
-          checkedTerritories.push(this.territories[index]._id);
-        }
-        return checkedTerritories;
-      },
-      [],
-    ));
-    this._operatorVisilibityService.update(territoriesIds).subscribe(
+    this._operatorVisilibityService.update(this.selectedTerritoryIds).subscribe(
       () => {
         this._toastr.success('Modifications prises en compte.');
       },
@@ -110,12 +157,16 @@ export class OperatorVisibilityTreeComponent extends DestroyObservable implement
 
   public checkAll($event: any): void {
     if ($event.checked) {
-      const formValues = Array(this.territories.length).fill(true);
-      this.territoriesFormArray.setValue(formValues);
+      // this.territories.forEach((ter) => (ter.selected = true));
+      this.selectedTerritoryIds = this.territories
+        .filter(isSelectableFilter)
+        .map((ter) => ter._id)
+        .filter((val, ind, self) => self.indexOf(val) === ind);
     } else {
-      const formValues = Array(this.territories.length).fill(false);
-      this.territoriesFormArray.setValue(formValues);
+      // this.territories.forEach((ter) => (ter.selected = false));
+      this.selectedTerritoryIds = [];
     }
+    this.visibilityFormControl.setValue(this.selectedTerritoryIds);
   }
 
   public showTerritory(id: number): boolean {
@@ -123,25 +174,21 @@ export class OperatorVisibilityTreeComponent extends DestroyObservable implement
   }
 
   public get countCheckedTerritories(): number {
-    return this.checkedTerritoryIds.length;
+    return this.selectedTerritoryIds.filter((val, ind, self) => self.indexOf(val) === ind).length;
   }
 
   private initVisibilityForm(): void {
-    this.visibilityFormGroup = this._fb.group({
-      territories: this._fb.array([]),
-    });
+    this.visibilityFormControl = this._fb.control([]);
   }
 
   private updateVisibilityForm(): void {
-    const territories = this.territories.sort((a, b) => a.name.localeCompare(b.name));
-    const formGroups = [];
-    const territoryIds = this.checkedTerritoryIds;
-    for (const territory of territories) {
-      formGroups.push(this._fb.control(territoryIds.indexOf(territory._id) !== -1));
-    }
-    this.visibilityFormGroup = this._fb.group({
-      territories: this._fb.array(formGroups),
-    });
+    if (!this.territories || !this.selectedTerritoryIds) return;
+
+    const territoryIds = this.selectedTerritoryIds;
+
+    this.territories.forEach((ter) => (ter.selected = ter.activable && territoryIds.indexOf(ter._id) !== -1));
+
+    this.visibilityFormControl.setValue(territoryIds);
   }
 
   private initSearchForm(): void {
