@@ -1,6 +1,7 @@
 export function integerMacro(): object {
   return {
     type: 'integer',
-    minimum: 1,
+    minimum: -2147483647,
+    maximum: 2147483647,
   };
 }
