@@ -210,7 +210,6 @@ export class HttpTransport implements TransportInterface {
           makeCall(
             'acquisition:cancel',
             {
-              ...req.body,
               journey_id: parseInt(req.params.id, 10),
             },
             { user, metadata: { req } },
@@ -234,7 +233,7 @@ export class HttpTransport implements TransportInterface {
         );
 
         const response = (await this.kernel.handle(
-          makeCall('acquisition:create', req.body, { user, metadata: { req } }),
+          makeCall('acquisition:create', { ...req.body }, { user, metadata: { req } }),
         )) as RPCResponseType;
 
         this.send(res, response);
@@ -248,7 +247,7 @@ export class HttpTransport implements TransportInterface {
       rateLimiter(),
       asyncHandler(async (req, res, next) => {
         const response = (await this.kernel.handle(
-          makeCall('trip:stats', {}, { user: { permissions: ['trip.stats'] } }),
+          makeCall('trip:stats', { tz: decodeURIComponent(req.query.tz) }, { user: { permissions: ['trip.stats'] } }),
         )) as RPCResponseType;
 
         if (!response || Array.isArray(response) || 'error' in response) {
@@ -423,61 +422,7 @@ export class HttpTransport implements TransportInterface {
     );
   }
 
-  // FIXME
-  // - add server authentication
-  // - block access to POST /certificates route
   private registerCertificateRoutes(): void {
-    /**
-     * The route that renders an HTML certificate based on params
-     * - only accessible by the printer with JWT authentication
-     * - requires access to public assets (images)
-     *
-     * TEST ME : http://localhost:8080/v2/certificates/render?identity=%2B
-     */
-    this.app.get(
-      '/v2/certificates/render/:uuid',
-      rateLimiter(),
-      async (req, res, next) => {
-        try {
-          // check the token generated by the download action
-          const decoded = (await this.tokenProvider.verify(req.headers.authorization.replace('Bearer ', ''), {
-            issuer: this.config.get('proxy.apiUrl'),
-          })) as { uuid: string; iat: number };
-
-          // 5s expiration
-          if (!decoded?.iat || decoded.iat > new Date().getTime() + 5000) {
-            throw new UnauthorizedException('Expired token');
-          }
-
-          // uuid must match
-          if (decoded?.uuid !== req.params.uuid) {
-            throw new UnauthorizedException('Wrong token');
-          }
-
-          next();
-        } catch (e) {
-          next(e);
-        }
-      },
-      asyncHandler(async (req, res, next) => {
-        const response = (await this.kernel.handle(
-          makeCall(
-            'certificate:render',
-            {
-              uuid: req.params.uuid,
-              token: get(req, 'headers.authorization', '').replace('Bearer ', ''),
-            },
-            { user: { permissions: ['certificate.render'] } },
-          ),
-        )) as RPCResponseType;
-
-        this.raw(res, get(response, 'result.data', response), { 'Content-type': 'text/html' });
-      }),
-    );
-
-    // public assets routes
-    this.app.use('/v2/certificates/assets', express.static('../../services/certificate/dist/assets'));
-
     /**
      * Public route to check a certificate
      */
@@ -498,20 +443,15 @@ export class HttpTransport implements TransportInterface {
      * - accessible with an application token
      * - uses /v2/certificates/render to capture the rendered certificate
      * - uses the remote printer to capture the rendered certificate
-     * - print a PDF/PNG returned back to the caller
+     * - print a PDF returned back to the caller
      */
     this.app.get(
-      '/v2/certificates/:type/:uuid/',
+      '/v2/certificates/pdf/:uuid/',
       rateLimiter(),
       asyncHandler(async (req, res, next) => {
-        const type = req.params.type.toLowerCase();
         const uuid = req.params.uuid.replace(/[^a-z0-9-]/gi, '').toLowerCase();
 
-        const call = makeCall(
-          'certificate:download',
-          { uuid, type },
-          { user: { permissions: ['certificate.download'] } },
-        );
+        const call = makeCall('certificate:download', { uuid }, { user: { permissions: ['certificate.download'] } });
         const response = (await this.kernel.handle(call)) as RPCResponseType;
 
         this.raw(res, get(response, 'result.body', response), get(response, 'result.headers', {}));
@@ -532,7 +472,11 @@ export class HttpTransport implements TransportInterface {
         const response = (await this.kernel.handle(
           makeCall(
             'certificate:create',
-            { ...req.body, operator_id: get(req, 'session.user.operator_id') },
+            {
+              tz: req.body.tz,
+              identity: req.body.identity,
+              operator_id: get(req, 'session.user.operator_id'),
+            },
             { user: get(req, 'session.user', null) },
           ),
         )) as RPCResponseType;
@@ -573,7 +517,7 @@ export class HttpTransport implements TransportInterface {
       '/honor',
       monHonorCertificateRateLimiter(),
       asyncHandler(async (req, res, next) => {
-        await this.kernel.handle(makeCall('honor:save', { ...req.body }, { channel: { service: 'proxy' } }));
+        await this.kernel.handle(makeCall('honor:save', { type: req.body.type }, { channel: { service: 'proxy' } }));
         res.status(201).header('Location', `${process.env.APP_APP_URL}/stats`).json({ saved: true });
       }),
     );
