@@ -1,4 +1,3 @@
-import path from 'path';
 import { serviceProvider, NewableType, ExtensionInterface } from '@ilos/common';
 import { ServiceProvider as AbstractServiceProvider } from '@ilos/core';
 import { PostgresConnection } from '@ilos/connection-postgres';
@@ -8,7 +7,7 @@ import { PermissionMiddleware } from '@pdc/provider-acl';
 import { DateProvider } from '@pdc/provider-date';
 import { QrcodeProvider } from '@pdc/provider-qrcode';
 import { CryptoProvider } from '@pdc/provider-crypto';
-import { PrinterProvider } from '@pdc/provider-printer';
+import { PdfCertProvider } from '@pdc/provider-pdfcert';
 import { TemplateExtension } from '@pdc/provider-template';
 
 import { config } from './config';
@@ -16,13 +15,11 @@ import { CertificatePgRepositoryProvider } from './providers/CertificatePgReposi
 import { IdentityPgRepositoryProvider } from './providers/IdentityPgRepositoryProvider';
 import { CarpoolPgRepositoryProvider } from './providers/CarpoolPgRepositoryProvider';
 import { TerritoryPgRepository } from './providers/TerritoryPgRepositoryProvider';
-import { RenderCertificateAction } from './actions/RenderCertificateAction';
 import { CreateCertificateAction } from './actions/CreateCertificateAction';
 import { FindCertificateAction } from './actions/FindCertificateAction';
 import { ListCertificateAction } from './actions/ListCertificateAction';
 import { DownloadCertificateAction } from './actions/DownloadCertificateAction';
 import { SeedCommand } from './commands/SeedCommand';
-import { binding as renderBinding } from './shared/certificate/render.schema';
 import { binding as createBinding } from './shared/certificate/create.schema';
 import { binding as findBinding } from './shared/certificate/find.schema';
 import { binding as downloadBinding } from './shared/certificate/download.schema';
@@ -38,27 +35,17 @@ import { binding as listBinding } from './shared/certificate/list.schema';
     IdentityPgRepositoryProvider,
     CarpoolPgRepositoryProvider,
     TerritoryPgRepository,
-    PrinterProvider,
+    PdfCertProvider,
   ],
-  validator: [renderBinding, createBinding, findBinding, downloadBinding, listBinding],
+  validator: [createBinding, findBinding, downloadBinding, listBinding],
   middlewares: [
     ['validate', ValidatorMiddleware],
     ['can', PermissionMiddleware],
     ['channel.service.only', ChannelServiceWhitelistMiddleware],
   ],
   connections: [[PostgresConnection, 'connections.postgres']],
-  handlers: [
-    RenderCertificateAction,
-    DownloadCertificateAction,
-    CreateCertificateAction,
-    FindCertificateAction,
-    ListCertificateAction,
-  ],
+  handlers: [DownloadCertificateAction, CreateCertificateAction, FindCertificateAction, ListCertificateAction],
   commands: [SeedCommand],
-  template: {
-    path: path.resolve(__dirname, 'templates').replace('/dist/', '/src/'),
-    meta: 'templates',
-  },
 })
 export class ServiceProvider extends AbstractServiceProvider {
   readonly extensions: NewableType<ExtensionInterface>[] = [ValidatorExtension, TemplateExtension];
