@@ -4,15 +4,15 @@ import { Action as AbstractAction } from '@ilos/core';
 import { CampaignRepositoryProviderInterfaceResolver } from '../interfaces/CampaignRepositoryProviderInterface';
 import { handlerConfig, ResultInterface } from '../shared/policy/templates.contract';
 import { alias } from '../shared/policy/templates.schema';
-import { ActionMiddleware } from '../shared/common/ActionMiddlewareInterface';
 
-@handler(handlerConfig)
-export class TemplatesCampaignAction extends AbstractAction {
-  public readonly middlewares: ActionMiddleware[] = [
+@handler({
+  ...handlerConfig,
+  middlewares: [
     ['can', ['incentive-campaign.templates']],
     ['validate', alias],
-  ];
-
+  ],
+})
+export class TemplatesCampaignAction extends AbstractAction {
   constructor(private campaignRepository: CampaignRepositoryProviderInterfaceResolver) {
     super();
   }
