@@ -9,6 +9,8 @@ import {
   RestrictionUxInterface,
   RetributionUxInterface,
 } from '~/core/interfaces/campaign/ux-format/campaign-ux.interface';
+import { CampaignReducedStats } from '../api-format/CampaignStats';
+import { cloneDeep } from 'lodash-es';
 
 export class CampaignUx {
   public _id: number;
@@ -19,6 +21,7 @@ export class CampaignUx {
   public end: Moment;
   public status: CampaignStatusEnum;
   public parent_id: number;
+  public state?: CampaignReducedStats;
   public unit: IncentiveUnitEnum;
   public filters: IncentiveFiltersUxInterface;
   public max_amount: number;
@@ -77,6 +80,11 @@ export class CampaignUx {
         staggered: null,
         insee_mode: null,
       },
+      state: {
+        amount: 0,
+        trip_subsidized: 0,
+        trip_excluded: 0,
+      },
     },
   ) {
     this._id = obj._id;
@@ -88,10 +96,11 @@ export class CampaignUx {
     this.parent_id = obj.parent_id;
     this.unit = obj.unit;
     this.filters = obj.filters;
-    this.restrictions = obj.restrictions;
-    this.retributions = obj.retributions;
+    this.restrictions = cloneDeep(obj.restrictions);
+    this.retributions = cloneDeep(obj.retributions);
     this.only_adult = obj.only_adult;
     this.ui_status = obj.ui_status;
+    this.state = obj.state;
 
     this.max_trips = obj.max_trips;
     this.max_amount = obj.max_amount;
