@@ -1,3 +1,4 @@
+import { removeAccents } from '../utils';
 export class Contact {
   firstname: string;
   lastname: string;
@@ -7,7 +8,7 @@ export class Contact {
   constructor(obj: { firstname: string; lastname: string; email: string; phone?: string }) {
     if (obj && obj.firstname) this.firstname = obj.firstname;
     if (obj && obj.lastname) this.lastname = obj.lastname;
-    if (obj && obj.email) this.email = obj.email;
+    if (obj && obj.email) this.email = removeAccents(obj.email);
     if (obj && obj.phone) this.phone = obj.phone;
   }
 
