@@ -13,6 +13,9 @@ import { ParamsInterface as DeleteParamsInterface } from '~/core/entities/api/sh
 import { ParamsInterface as CreateParamsInterface } from '~/core/entities/api/shared/policy/create.contract';
 import { ParamsInterface as PatchParamsInterface } from '~/core/entities/api/shared/policy/patch.contract';
 import { ParamsInterface as ListParamsInterface } from '~/core/entities/api/shared/policy/list.contract';
+import { AuthenticationService } from '~/core/services/authentication/authentication.service';
+import { CampaignReducedStats } from '~/core/entities/campaign/api-format/CampaignStats';
+import { CampaignInterface } from '~/core/entities/api/shared/policy/common/interfaces/CampaignInterface';
 
 @Injectable({
   providedIn: 'root',
@@ -27,13 +30,17 @@ export class CampaignApiService extends JsonRpcCrud<
   any,
   CreateParamsInterface
 > {
-  constructor(http: HttpClient, router: Router, activatedRoute: ActivatedRoute) {
+  constructor(http: HttpClient, router: Router, activatedRoute: ActivatedRoute, protected auth: AuthenticationService) {
     super(http, router, activatedRoute, 'campaign');
   }
 
   public launch(id: number): Observable<LaunchResultInterface> {
     const jsonRPCParam = new JsonRPCParam(`${this.method}:launch`, { _id: id });
     return this.callOne(jsonRPCParam).pipe(map((data) => data.data));
+  }
+
+  getById(id: number): Observable<Campaign> {
+    return this.get({ _id: id, territory_id: this.auth.user.territory_id } as any);
   }
 
   loadTemplates(): Observable<Campaign[]> {
@@ -48,5 +55,10 @@ export class CampaignApiService extends JsonRpcCrud<
         _id: params._id,
       })),
     );
+  }
+
+  simulate(campaign: CampaignInterface): Observable<CampaignReducedStats> {
+    const jsonRPCParam = new JsonRPCParam(`${this.method}:simulateOnPast`, { campaign });
+    return this.callOne(jsonRPCParam).pipe(map((data) => data.data));
   }
 }
