@@ -27,7 +27,6 @@ export class FilterService {
 
     const filter = new Filter({
       date: filterUx.date,
-      hour: filterUx.hour as any,
       days: filterUx.days,
       distance: filterUx.distance,
       ranks: filterUx.ranks,
@@ -37,9 +36,13 @@ export class FilterService {
       campaign_id: filterUx.campaignIds,
     });
 
-    if (filter.date.start && typeof filter.date.start === 'string')
+    if (filter.date.start && typeof filter.date.start === 'string') {
       filter.date.start = moment(filter.date.start) as any;
-    if (filter.date.end && typeof filter.date.end === 'string') filter.date.end = moment(filter.date.end) as any;
+    }
+
+    if (filter.date.end && typeof filter.date.end === 'string') {
+      filter.date.end = moment(filter.date.end) as any;
+    }
 
     // Apply default start date filter
 
@@ -49,24 +52,6 @@ export class FilterService {
       // set start at the beginning of the day and end at the end.
       if (filter.date.start) filter.date.start = (filter.date.start as any).startOf('day').toDate();
       if (filter.date.end) filter.date.end = (filter.date.end as any).endOf('day').toDate();
-    }
-
-    if (filter.hour.start === null && filter.hour.end === null) {
-      delete filter.hour;
-    } else {
-      // time zone hacky fix until ddb has timezones
-      if (filter.hour.start) {
-        filter.hour.start = moment().hours(filter.hour.start).utc().hours();
-      } else {
-        filter.hour.start = 0;
-      }
-
-      if (filter.hour.end) {
-        // time zone hacky fix until ddb has timezones
-        filter.hour.end = moment().hours(filter.hour.end).utc().hours();
-      } else {
-        filter.hour.end = 23;
-      }
     }
 
     // format distance to Number
