@@ -1,6 +1,6 @@
 export const environment = {
   production: false,
   name: 'dev',
-  apiUrl: 'https://pdc-dev.osc-fr1.scalingo.io/',
+  apiUrl: 'https://api-dev.covoiturage.beta.gouv.fr/',
   sentryDSN: 'https://7701d6142633431a9acecd5a7047be2c@sentry.data.gouv.fr/58',
 };
