@@ -12,7 +12,7 @@ export const logerror = {
     error_message: { oneOf: [{ type: 'null' }, { type: 'string', minLength: 0, maxLength: 4096 }] },
     error_code: { oneOf: [{ type: 'null' }, { type: 'string', minLength: 0, maxLength: 32 }] },
     error_line: { oneOf: [{ type: 'null' }, { type: 'integer', minimum: 1, maximum: 1000000 }] },
-    error_attempt: { type: 'integer' },
+    error_attempt: { type: 'integer', minimum: 1, maximum: 2147483647 },
     auth: { type: 'object', additionalProperties: true },
     headers: { type: 'object', additionalProperties: true },
     body: { type: 'object', additionalProperties: true },
