@@ -1,6 +1,7 @@
 import { AcquisitionInterface } from '../acquisition/common/interfaces/AcquisitionInterface';
+import { ParamsInterface as CrossCheckParamsInterface } from '../carpool/crosscheck.contract';
 
-export type ResultInterface = AcquisitionInterface | AcquisitionInterface[];
+export type ResultInterface = CrossCheckParamsInterface;
 export interface ParamsInterface extends AcquisitionInterface {}
 export const handlerConfig = {
   service: 'normalization',
